requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pytesseract
Pillow
requests
